push and pull works, no animations yet
